Created OrderAgnosticBinarySearch.java
We all know that the basic requirement for Binary Search is that the array must be sorted. But for the array to be sorted it can be sorted in any order, that is, either descending or ascending order. So, in order agnostic binary search the searching algorithm is the same as we do in binary search of time complexity O(log n), it's just that we make a check to determine whether the array is sorted in ascending or descending order.