Added Perfect Number program implementation in java
Added a Java program to check whether a number is a Perfect Number or Not A number whose sum of factors (excluding the number itself) is equal to the number is called a perfect number. In other words, if the sum of positive divisors (excluding the number itself) of a number equals the number itself is called a perfect number.