Create NumberGuessingGame.java
This is a number guessing game written in java language. In this game the computer randomly chooses a number and the user has to guess it. If the number is exactly same to the number generated by the computer then the user wins the game else the computer give the hint that weather the actual number is smaller or lager than the guessed number.

Please don't forget to change the class name from Main if it throws error.